MOVEONLY: Move utility functions from rpcwallet to wallet/rpc/util
